backend/app/requirements.txt
fastapi==0.115.5
openai==1.54.4
pydantic==2.9.2
python-dotenv==1.0.1
tree_sitter==0.23.2
tree_sitter_python==0.23.4
